Desktop/farmacia_v2/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.3.1
dj-database-url==2.1.0
